+++ src/python/futag-package/src/futag/sysmsg.py
@@ -6,7 +6,7 @@
 **   / __/   / /_/ /   / /   / ___ | / /_/ /    **
 **  /_/      \____/   /_/   /_/  |_| \____/     **
 **                                              **
-**     Fuzz target Automated Generator       **
+**     Fuzzing target Automated Generator       **
 **             a tool of ISP RAS                **
 **************************************************
 ** This module is for saving constants of Futag **