+++ CHANGELOG.md
@@ -5,6 +5,9 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## 20250824
+- Add support for Fuzzed Data Provider
+
 ## 20220716
 - Add modules preprocessor to Futag python-package
 - Fix README of Futag python-package